DBTAdapter/HCIHandler: Prefer using simple mutex/lock on CopyOnWrite (COW) for lock-free listener iteration and invocation.
Goal is to have a lock-free listener iteration and callback invocation to avoid any thinkable sideffects.

This CopyOnWrite (COW) implementation is not clean, as it still copies
the collection before iteration, as the entry is not a shared_ptr.
Hence would need refinement in this regard.

However, the copy and read is lock-free.
An experimental 1st step.