LFRingbuffer: Use std::lock(..) on operations waiting for either read and/or write operations to complete
std::lock(..) allows us to wait for either order of lock-able mutex in a list using try_lock.

Global operations like drop, clear, reset etc are unaware of the ringbuffer utilization and
need to wait for the global syncMulti* read and write locks in any order.

std::scoped_lock can't be used, as we utilize std::condition_variable.wait(..)
and don't want to 'mix and match' the locking type.