helgrind 'lock order': GATTHandler::exchangeMTU*: Called from ctor, no need to lock mtx_command
Helgrind would record the lock order DBTDevice::mtx_gattHandler (todo) and then mtx_command,
which only exists once at ctor.

Normally the order is mtx_command and if this leads to a disconnect,
adds DBTDevice::mtx__connect.

Since exchangeMTU* is not a public entry, rename it and pass the timeout value
for more clarity.