GATTHandler::exchangeMTU*: Handle accepted case of Opcode::ATT_ERROR_RSP w/ ErrorCode::UNSUPPORTED_REQUEST -> ATT_MTU
If the ATT_ERROR_RSP PDU is sent by the server
with the error code set to 'Request Not Supported',
the Attribute Opcode is not supported and the default MTU shall be used.

Add DBG_PRINT or ERR_PRINT for all cases.

Quering the MTU is quite a sensitive initial GATT operation,
as GATTHandler ctor will disconnect if failing.