BTDevice::connectLE: Allow explicit connect request to BLERandomAddressType [UNRESOLVABLE_PRIVAT and RESOLVABLE_PRIVAT]
Since we do not resolve the RESOLVABLE_PRIVAT using an IRK,
we simply set the HCILEPeerAddressType to RANDOM for all BDAddressType::BDADDR_LE_RANDOM peers.

HCILEOwnAddressType is always PUBLIC, as we do not support 'LE Set Privacy Mode' yet.

This at least seems to work OK for BLERandomAddressType::UNRESOLVABLE_PRIVAT.