Add BTGattChar::removeCharListener() and add convenience BTGattChar::disableIndicationNotification(); On the Java ..
On the Java side the implementation had to be a little different.

The wrapping BTGattCharListener from addCharListener() can't be repeated as it holds the
native instance handle.

Here addCharListener() needs to return the wrapping BTGattCharListener object,
which the user shall reference if intending a later BTGattChar::removeCharListener().

This avoid maintaining a map(BTGattChar::Listener -> BTGattCharListener )
and hence the user takes the costs if desired only.