Removed TinyB API artifacts, fix copyright notice. Add README.md 'Direct-BT Origins'
Removed remaining TinyB API artifacts:
- BTType.java and related search functions (class_type())
- Java's clone()
- BTManager's device/gatt functionality, only covers adapter.

+++

Direct-BT originated from TinyB.

Major TinyB autors were:
- Andrei Vasiliu <andrei.vasiliu@intel.com>
- Petre Eftime <petre.p.eftime@intel.com>

TinyB Copyright (c) 2016 Intel Corporation.

With the removal of TinyB from Direct-BT
we removed all substantial parts (if not all, i.e. > 99.9%)
of its copyright holder.

This change removes remaining source file Intel (c) notices.

This change does not disregard the author's original work,
on the contrary, we keep attribution about Direct-BT's history
in the README.md file 'Direct-BT Origins' section.

The Intel (c) notice and TinyB authors stay in the COPYING file.

We are grateful for their excellent work.

This change is to clarify authorship and copyright of Direct-BT's work,
as well as to not mislead association with the work of TinyB.