Trial: DBT[Client|Server]00: Have Client <-> Server handshake on success, allowing server to countdown servingConnectionsLeft to determine to stop or continue advertising
The TestDBTClientServer10 starts client and server instances then waits for a successful completion or a timeout will abort the test.

A successful completion requires both, client and server, to have determined this state
and stop discover and advertising respectively.

The server's processDisconnectedDevice() shall re-enable advertising in case
more connections are to be served.

For example a SMP failure would disconnect the client from the server
without a successful test completion.

Hence the server needs to determine a successful test,
which is now being signaled via a GATT command by the client.
Here the client writes a specific value to a GATT characteristic,
which is recognized by the server and returned to the client.
If done, client and server countdown their counter of tests left.

This renders the client server test more robust
as required for certain BT adapter setups w/ failing connections.
Here the test will continue until success or timeout.